+++ client/src/redux/actions/twitterActions.js
@@ -16,9 +16,6 @@ import {
   TWITTER_FILTER_ACCOUNTS,
   TWITTER_SET_TWEET_LIKE_STATUS,
   TWITTER_SET_TWEET_RETWEET_STATUS,
-  TWITTER_SET_TWEET_REPLIES,
-  TWITTER_RESET_TWEET_REPLIES,
-  TWITTER_SET_REPLIES_LOADING_STATUS,
 } from '../types';
 import { store } from '../store';
 import { v4 as uuidv4 } from 'uuid';
@@ -486,56 +483,4 @@ export const retweetTweet = ({
   } catch (err) {
     dispatch(handleError({ error: err }));
   }
-};
-
-export const fetchTweetReplies = ({
-  twitterUserId,
-  repliesToScreenName,
-  inReplyToStatusId,
-  sinceId,
-}) => async (dispatch) => {
-  await dispatch(setRepliesLoadingStatus({ isLoading: true }));
-  const response = await axiosInstance.get(
-    TwitterEndpoints.fetchTweetReplies({
-      twitterUserId,
-      repliesToScreenName,
-      inReplyToStatusId,
-      sinceId,
-    })
-  );
-  if (response.data.length === 0) {
-    await dispatch(setRepliesLoadingStatus({ isLoading: false }));
-    return;
-  }
-  const repliesArray = response.data.map((tweet) => tweet.id_str);
-  const tweetsObject = response.data.reduce(
-    (obj, tweet) => ({ ...obj, [tweet.id_str]: { ...tweet } }),
-    {}
-  );
-
-  await dispatch({
-    type: TWITTER_SET_TWEET_REPLIES,
-    payload: {
-      replies: tweetsObject,
-      repliesArray: repliesArray,
-      maxId: response.data[response.data.length - 1].id_str,
-      moreReplies: response.data.length === 30 ? true : false,
-    },
-  });
-  await dispatch(setRepliesLoadingStatus({ isLoading: false }));
-};
-
-export const resetReplies = () => (dispatch) => {
-  dispatch({
-    type: TWITTER_RESET_TWEET_REPLIES,
-  });
-};
-
-export const setRepliesLoadingStatus = ({ isLoading }) => (dispatch) => {
-  dispatch({
-    type: TWITTER_SET_REPLIES_LOADING_STATUS,
-    payload: {
-      isLoading,
-    },
-  });
 };
